pyproject.toml
 [build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ecommerce-recsys"
version = "0.1.0"
description = "E-commerce recommendation system"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Fuad Gurbanov", email = "fuad.gurbanove@gmail.com"},
]

dependencies = [
    "pydantic>=2.0.0",
    "confluent-kafka>=2.0.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]

dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]

select = ["E", "F", "I", "N"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = false
warn_required_dynamic_aliases = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
